tests/wpt/tests/shadow-dom/focus-navigation/menu/tentative/focus-menu-elements.html

<!DOCTYPE html>
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<link rel=help href=https://open-ui.org/components/menu.explainer>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src='../../resources/shadow-dom.js'></script>
<script src="../../resources/focus-utils.js"></script>

<button id=before>before</button>

<menubar>
  <menuitem id="A" command="toggle-menu" commandfor="more">More commands</menuitem>
  <menuitem id="B">Command B</menuitem>
  <menuitem id="C">Command C</menuitem>
</menubar>

<button id=middle>middle</button>

<menulist id="more">
 <menuitem id="A1">Command A1</menuitem>
 <menuitem id="A2">Command A2</menuitem>
</menulist>

<button id=after>after</button>

<style>
  menuitem {
    /* Make sure focus doesn't activate the submenu. */
    interest-delay: 1000s;
  }
</style>

<script>
promise_test(async () => {
  let elements = [
    'before',
    'A',
    'middle',
    'after',
  ];

  await assert_focus_navigation_bidirectional(elements);
}, 'Focus navigation should skip over other menuitems in menubars.');
</script>
